10-100x speedup for ScatterPlotItem

+++ examples/ScatterPlotSpeedTest.py
@@ -26,7 +26,7 @@ win.show()
 p = ui.plot
 
 data = np.random.normal(size=(50,500), scale=100)
-sizeArray = np.random.random(500) * 20.
+sizeArray = (np.random.random(500) * 20.).astype(int)
 ptr = 0
 lastTime = time()
 fps = None
